API test moved to separate directory
